+++ packages/eyes-sdk-core/lib/utils/takeDomSnapshot.js
@@ -13,7 +13,8 @@ const createFramesPaths = require('./createFramesPaths')
 
 const EXECUTION_TIMEOUT = 5 * 60 * 1000
 const POLL_TIMEOUT = 200
-const DEFAULT_CHUNK_BYTE_LENGTH = 262144000 // 250MB (could be 256MB but decide to leave a 6MB buffer)
+const DEFAULT_CHUNK_BYTE_LENGTH =
+  Number(process.env.APPLITOOLS_SCRIPT_RESULT_MAX_BYTE_LENGTH) || 262144000 // 250MB (could be 256MB but decide to leave a 6MB buffer)
 
 async function takeDomSnapshot(logger, driver, options = {}) {
   ArgumentGuard.notNull(logger, 'logger')