+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Change Log
 
 ## Unreleased
+
 - add support for SDKs that don't support traversing between frames (e.g., it doesn't implement parentContext)
 - republish to update vgc due to cyclic dependency
+- updated to axios@0.21.1 (from 0.19.2)
 
 ## 12.11.1 - 2021/1/7