+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add env var APPLITOOLS_SCRIPT_RESULT_MAX_BYTE_LENGTH
 - add support for SDKs that don't support traversing between frames (e.g., it doesn't implement parentContext)
 - republish to update vgc due to cyclic dependency
 - updated to axios@0.21.1 (from 0.19.2)